chore: add lib/event-types.ts, suppress Node 26 deprecation warning
Split PublicEvent/PublicWave types and lowestPrice out of lib/events.ts
into lib/event-types.ts to avoid pulling Prisma/pg into client bundles.
Set NODE_OPTIONS=--no-deprecation in test:e2e to silence module.register()
warning on Node 26.
